Movement and jumping now all works

+++ src/com/scs/joustgame/EntityFactory.java
@@ -29,7 +29,7 @@ public class EntityFactory {
 	}
 	
 
-	public AbstractEntity createPlayersAvatar(PlayerData player, IPlayerInput controller, int x, int y) {
+	public AbstractEntity createPlayersAvatar(PlayerData player, IPlayerInput controller, float x, float y) {
 		AbstractEntity e = new AbstractEntity("Player");
 
 		ImageComponent imageData = new ImageComponent("grey_box.png", 1, Settings.PLAYER_SIZE, Settings.PLAYER_SIZE);
